Clean up GitHub Actions workflow
We are working on consolidating all CI to GitHub Actions and while this
repo does have a workflow, it's also making use of an experiment around
devbox which did not continue. As such I am clearing out it's use of
devbox and going with more straight forward scripts, as well as updating
the matrix of Python versions we support.